requirements.txt
pandas
yfinance
ta
fastapi
uvicorn
google-cloud-firestore
pytest
